change unit prices to decimals

+++ db/migrate/20150915222609_add_columns_to_invoice_items.rb
@@ -2,6 +2,6 @@ class AddColumnsToInvoiceItems < ActiveRecord::Migration
   def change
     add_column :invoice_items, :quantity, :integer
 
-    add_column :invoice_items, :unit_price, :float
+    add_column :invoice_items, :unit_price, :integer
   end
 end
